// src/modules/log/enums/logEnum.ts
export enum OrderColumns {
  HOST = 'host',
  SERVICE = 'service',
  PID = 'pid',
  MESSAGE = 'message',
  DATETIME = 'datetime',
  TIME_EXECUTION = 'time_execution',
  TARGET = 'target',
}

export enum OrderBy {
  ASC = 'ASC',
  DESC = 'DESC',
}

export enum TableColumnsName {
  ID = 'id',
  HOST = 'host',
  SERVICE = 'service',
  PID = 'pid',
  MESSAGE = 'message',
  DATETIME = 'datetime',
  TIME_EXECUTION = 'timeExecution',
  TARGET = 'target',
  PREDICTED_LOG = 'predictedLog',
  VERIFIED_LOG = 'verifiedLog',
}

export enum TableColumnsLabel {
  ID = 'ID',
  HOST = 'Host',
  SERVICE = 'Service',
  PID = 'PID',
  MESSAGE = 'Message',
  DATETIME = 'Datetime',
  TIME_EXECUTION = 'Time execution',
  PREDICTED_LOG = 'Predicted value',
  VERIFIED_LOG = 'Verify',
}

export enum TableRowKey {
  ID = 'id'
}